utilities: simplify and fix conversion of relative lengths to pt
`convert-length-to-pt` had a couple of issues regarding relative
lengths.

When the `length` component has an em part, there would be an error
because Typst does not allow comparing `length`s with an em part with
those that do not.

It also did not account for ratios with a fraction (e.g., 1.5%).

This commit fixes those issues and, at the same time, simplifies the
implementation to re-use the other conversion functions for better
consistency.