utilities: fix issues related to conversion of lengths to pt (#84)
* utilities: fix conversion of negative `length` lengths

The `convert-length-to-pt` function did not properly handle converting
negative `length`s to pt. They would get converted to 0pt because the
line drawn for measurement would (I presume) have a width of 0pt
when given a negative `length`.

* utilities: factor out pt conversions into functions

* utilities: simplify and fix conversion of relative lengths to pt

`convert-length-to-pt` had a couple of issues regarding relative
lengths.

When the `length` component has an em part, there would be an error
because Typst does not allow comparing `length`s with an em part with
those that do not.

It also did not account for ratios with a fraction (e.g., 1.5%).

This commit fixes those issues and, at the same time, simplifies the
implementation to re-use the other conversion functions for better
consistency.

* utilities: convert `convert-length-to-pt` variables to kebab case

* utilities: convert `length`s to pt without losing precision

Unfortunately, `repr` rounds the `pt` component of `length`s to 2
decimal places, which results in a loss of precision.

This commit rewrites the conversion of `length`s to `pt` without
relying on the numerical values in the `repr`. We instead draw and
measure lines.

* utilities: convert `relative` lengths to pt without losing precision

Typst rounds all length components, except `em` components, to 2 decimal
places for their `repr`, which results in a loss of precision.

This commit resolves the loss of precision. Some aspects of Typst's
behaviour are important:
- The `repr` of `em` is lossless.
- The "draw and measure a line" technique applied on `relative` lengths
  returns the length of the `length` component only (the ratio component
  is ignored due to limitations of the technique).

* utilities: assert `page-size` and `frac-total` are purely pt lengths

* tablex-test: add tests for `convert-length-to-pt`

* tablex-test: add tests for line expansion

* tablex-test: check return type of conversion to pt

* tablex-test: test precision of conversion to pt

* add issue number to tablex-test.typ

* pdf will be in separate commit